backends: stop escaping backslashes in preprocessor defines
We give some special treatment to preprocessor defines when assembling
the compiler arguments: instead of passing them as-is, we escape any
backslash characters. This is done ever since aca93df18 (backend: There
is no need for 'compiler-specific escaping' anymore, 2019-04-30), which
is part of a bigger refactoring of the Ninja backend so that Ninja
itself handles shell quoting of arguments. Quoting the commit message:

    Now that all command-line escaping for ninja is dealt with in the
    ninja backend, escape_extra_args() shouldn't need to do anything.

    But tests of existing behaviour rely on all backslashes in defines
    being C escaped: This means that Windows-style paths including
    backslashes can be safely used, but makes it impossible to have a
    define containing a C escape.

So the original intent of the quoting was to make it easier to work with
Windows-style paths. The commit also mentions that this behaviour is not
needed anymore now that we handle quoting in Ninja itself, but that we
cannot alter existing behaviour because tests rely on it.

What's curious though is that before the mentioned commit, we only used
to quote on Windows. And that kind of makes sense, as the usecase was
Windows-style paths. It's also arguable that having different quoting
behaviour on different platforms would cause confusion.

But there's a significant issue with the current behaviour: it is
impossible for the user of Meson to use backslashes to perform manual
escaping. As such, passing string-like preprocessor defines to Meson
is impossible where the string is supposed to contain C escapes itself.

Due to this inherent limitation, Muon has decided to not follow Meson's
behaviour and instead opts to not escape backslash characters [1]:

    Meson replaces \ with \\ in compiler defines. This is legacy
    behavior that prevents you from using things like C escapes (e.g.
    \n) in compiler defines, at the benefit of making it easier to use
    windows paths.

It is unfortunate to have such diverging behaviour, but the case that
Muon makes is quite sensible: is it simply impossible to get C escapes
into preprocessor defines, whereas it would be possible to manually
quote Windows-style paths as required.

Fix this broken usecase by not quoting backslash characters anymore.
This enables C escapes in preprocessor defines and realigns the
behaviour of Meson and Muon.

[1]: https://docs.muon.build/differences.html#backslash-escaping-in-compiler-defines

Signed-off-by: Patrick Steinhardt <ps@pks.im>